refactor: rename implementation.ts files to class-name convention
Follow codebase convention where implementation files are named after
the class they contain (e.g., ExecuteSync.ts, not implementation.ts).
