+++ packages/api-sync-ddb-to-opensearch/src/features/DdbToOpenSearchHandler/DdbToOpenSearchHandler.ts
