+++ packages/api-sync-to-opensearch/src/features/ExecuteSyncWithRetry/feature.ts
@@ -1,5 +1,5 @@
 import { createFeature } from "@webiny/feature/api";
-import { ExecuteSyncWithRetryImplementation } from "./implementation.js";
+import { ExecuteSyncWithRetryImplementation } from "./ExecuteSyncWithRetry.js";
 
 export const ExecuteSyncWithRetryFeature = createFeature({
     name: "sync.executeSyncWithRetry",
