+++ packages/api-sync-to-opensearch/src/features/SynchronizationBuilder/feature.ts
@@ -1,5 +1,5 @@
 import { createFeature } from "@webiny/feature/api";
-import { SynchronizationBuilderImplementation } from "./implementation.js";
+import { SynchronizationBuilderImplementation } from "./SynchronizationBuilder.js";
 
 export const SynchronizationBuilderFeature = createFeature({
     name: "sync.synchronizationBuilder",
